Some fixes. Another attempt to fix server patching. DPI setting was removed from preferences. The recommended way to change dpi is to use `-dpi N` option of X server. Fix for captureless Dex touchpad handler. Now it should work.

+++ app/src/main/java/com/termux/x11/input/InputEventSender.java
@@ -67,7 +67,7 @@ public final class InputEventSender {
     }
 
     public void sendMouseWheelEvent(float distanceX, float distanceY) {
-        mInjector.sendMouseWheelEvent((int) distanceX, (int) distanceY);
+        mInjector.sendMouseWheelEvent(distanceX, distanceY);
     }
 
     public void sendReverseMouseWheelEvent(float distanceX, float distanceY) {
